pyproject.toml
[project]
name = "sorti"
version = "1.0.0"
description = "IMAP email processor for automatic document tagging"
authors = [
    { name = "Dennis", email = "1451140+karaktaka@users.noreply.github.com" }
]
dependencies = [
    "pyyaml>=6.0.1",
]
requires-python = ">=3.13"

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests"]

[dependency-groups]
test = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
]
